Second commit

+++ src/main/java/com/example/projecttaskmanagement/repository/UserRepository.java
@@ -4,13 +4,11 @@ import com.example.projecttaskmanagement.entity.User;
 
 import java.util.List;
 
-
-
 public interface UserRepository {
-    User findById(Long id);
+    User findById(int id);
     List<User> findAll();
     void save(User user);
     void update(User user);
-    void delete(Long id);
-    List<User> findAllByTaskId(Long taskId);
+    void delete(int id);
+    List<User> findAllByTaskId(int taskId);
 }
